Fixed compiler warnings:
DDASLLogger.m:93:20: Implicit conversion changes signedness: 'int' to 'size_t' (aka 'unsigned long')
DDASLLogCapture.m:132:27: Implicit conversion changes signedness: 'int' to 'unsigned long'
DDASLLogCapture.m:149:69: Implicit conversion changes signedness: 'int' to 'DDLogFlag' (aka 'enum DDLogFlag')
DDASLLogCapture.m:174:48: Implicit conversion changes signedness: '__darwin_time_t' (aka 'long') to 'unsigned long long'
DDASLLogCapture.m:215:34: Implicit conversion changes signedness: 'long long' to 'unsigned long long'
DDFileLogger.m:765:65: Implicit conversion changes signedness: 'uint64_t' (aka 'unsigned long long') to 'int64_t' (aka 'long long')
DDFileLogger.m:955:21: Implicit conversion changes signedness: 'int' to 'uintptr_t' (aka 'unsigned long')
DDASLLogCapture.m:48:75: '__MAC_OS_X_VERSION_MIN_REQUIRED' is not defined, evaluates to 0
DDASLLogCapture.m:102:26: 'TARGET_SIMULATOR' is not defined, evaluates to 0

+++ Classes/DDASLLogger.m
@@ -90,7 +90,7 @@ static DDASLLogger *sharedInstance;
 
         char readUIDString[16];
 #ifndef NS_BLOCK_ASSERTIONS
-        size_t l = snprintf(readUIDString, sizeof(readUIDString), "%d", readUID);
+        size_t l = (size_t)snprintf(readUIDString, sizeof(readUIDString), "%d", readUID);
 #else
         snprintf(readUIDString, sizeof(readUIDString), "%d", readUID);
 #endif
